requirements.txt
Django>=5.0
djangorestframework
drf-spectacular
requests
psycopg2-binary
djangorestframework-simplejwt
beautifulsoup4